token-refresh-service/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
